WIP: Some meaningful output

+++ src/Tools/dotnet-trace-hosted/Services/IMonitorService.cs
@@ -4,7 +4,7 @@ namespace HostedTrace
 {
     public interface IMonitorService
     {
-        Task<int> StartMonitorAsync(int processId);
+        Task<ulong> StartMonitorAsync(int processId);
         Task StopMonitor(string processId, int sessionId);
     }
 }
